tidy up after rebase
