+++ lading/Cargo.toml
@@ -55,14 +55,9 @@ proptest-derive = "0.3.0"
 
 [features]
 default = []
-dogstatsd_perf = []
 
 [lib]
 doctest = false
 
 [[bin]]
 name = "lading"
-
-[[bin]]
-name = "lading_rev"
-required-features = ["dogstatsd_perf"]
