--- lading/src/bin/lading_rev.rs
@@ -1,27 +1,0 @@
-/// A program whose sole purpose is to generate 1Gb of DogStatsD load and then
-/// exit. It's like having criterion benchmarks without having to split
-/// `lading::payload` into a crate.
-///
-/// I'm not sure how this will interact with the release process. We probably
-/// should disable this binary at some point.
-use lading::payload::dogstatsd;
-use rand::{rngs::SmallRng, SeedableRng};
-
-use std::io::{BufWriter, Write};
-
-fn main() {
-    let seed: u64 = 19690616;
-    let mut rng = SmallRng::seed_from_u64(seed);
-    let dg = dogstatsd::DogStatsD::default(&mut rng);
-
-    // NOTE we lock stdout and wrap it in a buffered writer to avoid needing to
-    // reacquire the stdio lock each write and, also, to elide as many writes as
-    // possible.
-    let stdout = std::io::stdout();
-    let mut fp = BufWriter::with_capacity(1_000_000, stdout.lock());
-    for _ in 0..1_000 {
-        let member = dg.generate(&mut rng);
-        writeln!(fp, "{member}").unwrap();
-    }
-    fp.flush().unwrap();
-}
